# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Movie API
  version: 1.0.0
servers:
- url: https://localhost:8080/api/v1
...
paths:
...
  /customers:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Customer'
                type: array
          description: OK
      summary: List all customers
      tags:
      - Customers
      x-accepts: application/json
      x-tags:
      - tag: Customers
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: Created
      summary: Register a new customer
      tags:
      - Customers
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Customers
  /customers/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_customers__id__get_404_response'
          description: The specified customer was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_movies__id__put_500_response'
          description: An internal server error occurred
      summary: Delete a customer account (only admin users can perform this action)
      tags:
      - Customers
      x-accepts: application/json
      x-tags:
      - tag: Customers
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_customers__id__get_404_response'
          description: The specified customer was not found
      summary: Get details of a specific customer
      tags:
      - Customers
      x-accepts: application/json
      x-tags:
      - tag: Customers
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_customers__id__get_404_response'
          description: The specified customer was not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_customers__id__put_409_response'
          description: A customer with the same email already exists
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_movies__id__put_500_response'
          description: An internal server error occurred
      summary: Update details of a specific customer
      tags:
      - Customers
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Customers
...
components:
...
  schemas:
...
    Customer:
      example:
        password: password
        phone: phone
        customerId: 0
        name: name
        membershipStatus: Basic
        email: email
      properties:
        customerId:
          description: Unique identifier for the customer
          format: int64
          type: integer
        name:
          description: Customer's full name
          type: string
        email:
          description: Customer's email address
          format: email
          type: string
        phone:
          description: Customer's phone number
          type: string
        password:
          description: Customer's password (should be encrypted)
          format: password
          type: string
        membershipStatus:
          description: Customer's membership status
          enum:
          - Basic
          - Premium
          type: string
      type: object
...
    _movies__id__put_500_response:
      properties:
        message:
          description: Generic error message
          type: string
      type: object
...
    _customers__id__get_404_response:
      properties:
        message:
          description: Error message indicating customer not found
          type: string
      type: object
    _customers__id__put_409_response:
      properties:
        message:
          description: Error message indicating duplicate customer email
          type: string
      type: object
...